+++ env.example
@@ -5,9 +5,9 @@ TOKEN=YOUR_BOT_TOKEN
 BOT_USERNAME=@your_username_bot
 
 # Filepaths
-SAVED_DATES_FILEPATH=data/saved_dates.pkl
-CHAT_IDS_FILEPATH=data/chat_ids.pkl
+SAVED_DATES_FILEPATH=saved_dates.pkl
+CHAT_IDS_FILEPATH=chat_ids.pkl
 
 # Serie A Calendar
 SERIE_A_CALENDAR_URL=https://fixturedownload.com/download/serie-a-2023-WEuropeStandardTime.csv
-SERIE_A_CALENDAR_PATH=data/seriea_calendar.csv
+SERIE_A_CALENDAR_PATH=seriea_calendar.csv
