+++ fantaformazionireminder/src/constants/notifications_interval.py
